requirements.txt
pyyaml==6.0.2
argparse==1.4.0
setuptools==80.4.0
build==1.2.2.post1